# api/openapi.yaml
openapi: 3.1.0
info:
  title: Halo management API
  version: 0.1.0
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  description: |
    The JSON API behind the Halo console and account portal. Anything the web interface does, a script can do with the same requests.

    **Authentication.** Send either the `halo_session` cookie of a signed-in person, or an API key of a service account as `Authorization: Bearer hlk_…`. API keys work only when they have the `api` scope, act with the roles of their service account, and are ignored on the personal routes under `/me` and on the sign-in routes under `/auth`. The personal routes under `/me/` (with a trailing segment, not `/me` itself) also accept an access token that `halo login` obtained for the built-in Halo CLI application.

    **Roles.** Each operation states the roles it requires. A global administrator passes every role check. "Any administrator" means security, user, helpdesk or application administrator, or auditor.

    **Errors.** Every error is `{"error": {"code": "ERR_…", "message": "…"}}`. Match on `code`; show `message` to people. Besides the codes listed per operation, any request can fail with `500` and `ERR_INTERNAL`, and unknown routes answer `404` with `ERR_NO_ROUTE`. Requests with a method other than GET, HEAD or OPTIONS that come from another website get `403` with `ERR_CROSS_SITE`.

...
servers:
  - url: /api/v1
security:
  - session: []
  - apiKey: []
...
paths:
...
  /users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List every person
      description: 'Roles: any administrator. Service accounts are not included.'
      responses:
        '200':
          description: Every person, sorted by name.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/User'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      tags: [Users]
      operationId: inviteUser
      summary: Invite a person
      description: 'Roles: user administrator. Creates the account with status `invited` and a setup link, and emails the link when SMTP is configured.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, name]
              properties:
                email: {type: string, format: email, maxLength: 254}
                name: {type: string, maxLength: 200}
                title: {type: string, maxLength: 200}
                department: {type: string, maxLength: 200}
                location: {type: string, maxLength: 200}
                managerId: {type: string, description: Id of another person.}
                groupIds:
                  type: array
                  description: Assigned groups only.
                  items: {type: string}
      responses:
        '201':
          description: The new person and their setup link.
          content:
            application/json:
              schema:
                type: object
                required: [user, enrollUrl, expiresAt, emailed]
                properties:
                  user: {$ref: '#/components/schemas/User'}
                  enrollUrl: {type: string, format: uri}
                  expiresAt: {type: string, format: date-time}
                  emailed: {type: boolean, description: True when Halo also emailed the link.}
        '400': {$ref: '#/components/responses/BadJSON'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Invalid'}
  /users/{id}:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    get:
      tags: [Users]
      operationId: getUser
      summary: Get one person
      description: 'Roles: any administrator.'
      responses:
        '200':
          description: The person.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/User'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      tags: [Users]
      operationId: updateUser
      summary: Change a person's profile
      description: 'Roles: user or helpdesk administrator. Helpdesk administrators can change title, department and location only. Only a global administrator can change a person who holds an administrator role. Send only the fields to change; an empty `managerId` removes the manager.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: {type: string, format: email}
                name: {type: string, maxLength: 200}
                title: {type: string, maxLength: 200}
                department: {type: string, maxLength: 200}
                location: {type: string, maxLength: 200}
                managerId: {type: string}
      responses:
        '200':
          description: The updated person.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/User'}
        '400': {$ref: '#/components/responses/BadJSON'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Invalid'}
...
  /groups:
    get:
      tags: [Groups]
      operationId: listGroups
      summary: List every group with its member count
      description: 'Roles: any administrator.'
      responses:
        '200':
          description: The groups.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Group'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      tags: [Groups]
      operationId: createGroup
      summary: Create a group
      description: 'Roles: user administrator.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, kind]
              properties:
                name: {type: string, maxLength: 200}
                description: {type: string, maxLength: 2000}
                kind: {type: string, enum: [assigned, dynamic]}
                rule: {type: string, description: Required for dynamic groups., examples: ['user.department == "Engineering"']}
      responses:
        '201':
          description: The group.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Group'}
        '400': {$ref: '#/components/responses/BadJSON'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Invalid'}
  /groups/{id}:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    get:
      tags: [Groups]
      operationId: getGroup
      summary: Get one group
      description: 'Roles: any administrator.'
      responses:
        '200':
          description: The group.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Group'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      tags: [Groups]
      operationId: updateGroup
      summary: Rename a group, or change its description or rule
      description: 'Roles: user administrator. The kind cannot change.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: {type: string, maxLength: 200}
                description: {type: string, maxLength: 2000}
                kind: {type: string, enum: [assigned, dynamic], description: Accepted only when it equals the current kind.}
                rule: {type: string, description: Dynamic groups only.}
      responses:
        '200':
          description: The updated group.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Group'}
        '400': {$ref: '#/components/responses/BadJSON'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Invalid'}
    delete:
      tags: [Groups]
      operationId: deleteGroup
      summary: Delete a group
      description: 'Roles: user administrator. Refused with `409`, naming each one, while an application, access package, policy, identity provider (for accounts created on first sign-in), open access review or lifecycle rule still uses the group. Completed access reviews are kept with the group''s name.'
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /groups/{id}/members:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    get:
      tags: [Groups]
      operationId: listGroupMembers
      summary: List a group's members, or preview a rule
      description: 'Roles: any administrator. Deprovisioned people are left out.'
      parameters:
        - name: rule
          in: query
          description: Dynamic groups only. Returns who would match this rule instead of the saved one.
          schema: {type: string}
      responses:
        '200':
          description: The members.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/User'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/Invalid'}
    post:
      tags: [Groups]
      operationId: addGroupMember
      summary: Add a person to an assigned group
      description: 'Roles: user administrator. Adding someone who is already a member succeeds.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId: {type: string}
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '400': {$ref: '#/components/responses/BadJSON'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/DynamicGroup'}
        '422': {$ref: '#/components/responses/Invalid'}
  /groups/{id}/members/{userId}:
    parameters:
      - {$ref: '#/components/parameters/Id'}
      - {$ref: '#/components/parameters/UserId'}
    delete:
      tags: [Groups]
      operationId: removeGroupMember
      summary: Remove a person from an assigned group
      description: 'Roles: user administrator.'
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/DynamicGroup'}
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: halo_session
      description: The session cookie Halo sets when a person signs in.
    apiKey:
      type: http
      scheme: bearer
      description: An API key of a service account, starting with `hlk_`, with the `api` scope. Requests act with the service account's roles.
    cliAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: An access token that `halo login` obtained for the built-in Halo CLI application. Accepted only on routes under `/me/`.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: {type: string}
    UserId:
      name: userId
      in: path
      required: true
      schema: {type: string, examples: [usr_01J9Z8Q4M2N5P7R9T1V3X5Z7B9]}
...
  responses:
    NoContent:
      description: Done. No body.
    BadJSON:
      description: '`ERR_INVALID_JSON`: the body is not valid JSON, is larger than 1 MiB, or has an unknown field.'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Unauthenticated:
      description: '`ERR_UNAUTHENTICATED`: no valid session or key.'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Forbidden:
      description: '`ERR_FORBIDDEN`: your roles do not allow this. `ERR_CROSS_SITE` on changes from another website.'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
    NotFound:
      description: '`ERR_NOT_FOUND`: the item does not exist.'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Conflict:
      description: '`ERR_CONFLICT`: the change conflicts with the current state, such as a name that is already used.'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    DynamicGroup:
      description: '`ERR_CONFLICT`: the group is dynamic, so its members come from its rule.'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Invalid:
      description: '`ERR_INVALID`: a field is missing or invalid. The message names it.'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string, examples: [ERR_NOT_FOUND]}
            message: {type: string, examples: [This user was not found. It may have been deleted.]}
...
    Method:
      type: object
      required: [id, kind, label, addedAt, lastUsedAt]
      properties:
        id: {type: string, description: '`recovery-codes` for the set of recovery codes.'}
        kind: {type: string, enum: [passkey, security-key, totp, recovery-codes]}
        label: {type: string, examples: [Passkey · Safari on iOS]}
        addedAt: {type: string, format: date-time}
        lastUsedAt: {type: [string, 'null'], format: date-time}
        detail: {type: string, examples: [8 of 10 remaining]}
    User:
      type: object
      required: [id, email, emailVerified, name, title, department, location, status, strength, methods, roles, groupIds, appIds, managerId, source, createdAt, lastSignInAt]
      properties:
        id: {type: string, examples: [usr_01J9Z8Q4M2N5P7R9T1V3X5Z7B9]}
        email: {type: string, format: email}
        emailVerified: {type: boolean, description: 'True once the person redeemed a magic link, finished setup from a link Halo emailed, or signed in through an identity provider that verified the same address. Changing the email clears it. With the `email` scope, ID tokens and userinfo carry it as `email_verified`.'}
        name: {type: string}
        title: {type: string}
        department: {type: string}
        location: {type: string}
        status: {type: string, enum: [invited, active, suspended, deprovisioned]}
        strength: {type: string, enum: [phishing-resistant, multi-factor, single-factor]}
        methods:
          type: array
          items: {$ref: '#/components/schemas/Method'}
        roles:
          type: array
          description: Role names, such as `Global administrator`. Assign roles with their keys.
          items: {type: string}
        groupIds:
          type: array
          description: Assigned and rule-based groups.
          items: {type: string}
        appIds:
          type: array
          description: Applications the person is assigned to through their groups.
          items: {type: string}
        managerId: {type: [string, 'null']}
        source: {type: string, examples: [Halo, 'SCIM · Okta', Google]}
        createdAt: {type: string, format: date-time}
        lastSignInAt: {type: [string, 'null'], format: date-time}
    Group:
      type: object
      required: [id, name, description, kind, memberCount, source, createdAt]
      properties:
        id: {type: string, examples: [grp_01J9Z8Q4M2N5P7R9T1V3X5Z7B9]}
        name: {type: string}
        description: {type: string}
        kind: {type: string, enum: [assigned, dynamic]}
        rule: {type: string, description: Dynamic groups only., examples: ['user.department == "Engineering"']}
        memberCount: {type: integer}
        source: {type: string}
        createdAt: {type: string, format: date-time}
...